+++ package.json
@@ -4,14 +4,17 @@
   "description": "Simple Chat",
   "private": true,
   "scripts": {
-    "dev": "nuxt",
-    "build": "nuxt build",
+    "dev": "nuxt-ts",
+    "build": "nuxt-ts build",
+    "generate": "nuxt-ts generate",
+    "start": "nuxt-ts start",
     "lint": "eslint --color --ext .tsx,.ts,.js,.vue .",
     "lint-fix": "eslint --fix --color --ext .tsx,.ts,.js,.vue .",
     "test": "jest"
   },
   "dependencies": {
     "@microsoft/signalr": "^3.0.1",
+    "@nuxt/typescript-runtime": "^0.3.8",
     "axios": "^0.19.2",
     "consola": "^2.11.3",
     "cross-env": "^7.0.0",
@@ -25,13 +28,12 @@
     "vuetify": "^2.2.11"
   },
   "devDependencies": {
-    "@nuxt/types": "^0.6.1",
     "@nuxt/typescript-build": "^0.5.6",
-    "@nuxt/typescript-runtime": "^0.3.8",
     "@nuxtjs/eslint-config-typescript": "^1.0.0",
     "@nuxtjs/vuetify": "^1.10.3",
     "@types/express": "^4.17.2",
     "@types/jest": "^25.1.2",
+    "@types/node": "^13.7.1",
     "@types/uuid": "^3.4.7",
     "@vue/test-utils": "^1.0.0-beta.31",
     "eslint": "^6.8.0",
