+++ src/store/index.ts
@@ -1,0 +1,5 @@
+const state = () => ({
+  authenticated: false
+})
+
+export { state }
