--- src/typings/shims-tsx.d.ts
@@ -1,11 +1,0 @@
-import Vue, { VNode } from 'vue'
-
-declare global {
-  namespace JSX {
-    interface Element extends VNode { }
-    interface ElementClass extends Vue { }
-    interface IntrinsicElements {
-      [elem: string]: any
-    }
-  }
-}
