--- src/typings/shims-vue.d.ts
@@ -1,14 +1,0 @@
-import Vue from 'vue'
-import VueRouter,{Route} from 'vue-router'
-
-declare module "*.vue" {
-  import Vue from "vue"
-  export default Vue
-}
-// Expansion
-declare module 'vue/types/vue' {
-  interface Vue {
-    $router: VueRouter,
-    $route: Route
-  }
-}
