+++ src/typings/vue-router-shim.d.ts
@@ -1,0 +1,8 @@
+import VueRouter, { Route } from 'vue-router'
+
+declare module 'vue/types/vue' {
+    interface Vue {
+        $router: VueRouter,
+        $route: Route
+    }
+}
