+++ src/typings/vue-shim.d.ts
@@ -1,0 +1,4 @@
+declare module "*.vue" {
+  import Vue from 'vue'
+  export default Vue
+}
