+++ tsconfig.json
@@ -1,25 +1,29 @@
 {
     "compilerOptions": {
-        "target": "es5",
-        "module": "es2015",
-        "strict": true,
+        "target": "es2018",
+        "module": "esnext",
         "moduleResolution": "node",
+        "lib": [
+            "esnext",
+            "dom"
+        ],
         "experimentalDecorators": true,
         "esModuleInterop": true,
+        "allowJs": true,
         "allowSyntheticDefaultImports": true,
         "sourceMap": true,
+        "strict": true,
+        "noEmit": true,
         "baseUrl": ".",
-        "types": [
-            "jest"
-        ],
         "paths": {
             "@/*": [
                 "src/*"
             ]
         },
-        "lib": [
-            "es2015",
-            "dom"
+        "types": [
+            "@types/node",
+            "@nuxt/types",
+            "jest"
         ]
     },
     "include": [
